Move topic, partition, producer, consumer into consumer mod
