+++ src/cli/src/consumer/consume/record_format.rs
