+++ src/cli/src/consumer/mod.rs
@@ -1,0 +1,5 @@
+pub mod error;
+pub mod topic;
+pub mod consume;
+pub mod produce;
+pub mod partition;
