+++ src/cli/src/consumer/partition/list.rs
