+++ src/cli/src/consumer/topic/delete.rs
