+++ src/cli/src/consumer/topic/describe.rs
