+++ src/cli/src/consumer/topic/list.rs
