+++ src/cli/src/error.rs
@@ -1,11 +1,13 @@
 use std::io::Error as IoError;
 
 use fluvio::FluvioError;
-use fluvio_extension_common::output::OutputError;
 use fluvio_cluster::cli::ClusterCliError;
-use crate::common::target::TargetError;
 use fluvio_sc_schema::ApiError;
 use fluvio_sc_schema::errors::ErrorCode;
+use fluvio_extension_common::output::OutputError;
+
+use crate::common::target::TargetError;
+use crate::consumer::error::ConsumerError;
 
 pub type Result<T> = std::result::Result<T, CliError>;
 
@@ -37,28 +39,6 @@ pub enum CliError {
     InvalidArg(String),
     #[error("Unknown error: {0}")]
     Other(String),
-}
-
-#[derive(thiserror::Error, Debug)]
-pub enum ConsumerError {
-    #[error(transparent)]
-    IoError(#[from] IoError),
-    #[error(transparent)]
-    OutputError(#[from] OutputError),
-    #[error("Fluvio client error")]
-    ClientError(#[from] FluvioError),
-    #[error("Invalid argument: {0}")]
-    InvalidArg(String),
-    #[error("Error finding executable")]
-    WhichError(#[from] which::Error),
-    #[error("Unknown error: {0}")]
-    Other(String),
-}
-
-impl ConsumerError {
-    pub fn invalid_arg<M: Into<String>>(reason: M) -> Self {
-        Self::InvalidArg(reason.into())
-    }
 }
 
 #[derive(thiserror::Error, Debug)]
